Making directory navigation work and changing tests

+++ src/_align.c
@@ -39,7 +39,7 @@ void *align_calloc(size_t count, size_t nbytes, const char *file, int line) {
                 Except_raise(&Mem_Failed, file, line);
         }
 
-    log_dbg("%p calloc %li bytes", ptr, nbytes);
+    log_dbg("%p calloc %lu bytes", ptr, (unsigned long) nbytes);
     return ptr;
 }
 
@@ -66,7 +66,7 @@ void *align_realloc(void *ptr, size_t nbytes, const char *file, int line) {
                 Except_raise(&Mem_Failed, file, line);
         }
 
-    log_dbg("%p realloc %li bytes", ptr, nbytes);
+    log_dbg("%p realloc %lu bytes", ptr, (unsigned long) nbytes);
     return ptr;
 }
